+++ planning/mapless_architecture/autoware_local_map_provider/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>local_map_provider</name>
+  <name>autoware_local_map_provider</name>
   <version>0.0.1</version>
   <description>local_map_provider</description>
   <maintainer email="simon.eisenmann@driveblocks.ai">driveblocks</maintainer>
