+++ planning/mapless_architecture/autoware_local_mission_planner/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mission_planner</name>
+  <name>autoware_local_mission_planner</name>
   <version>0.0.1</version>
   <description>Mission Planner</description>
   <maintainer email="simon.eisenmann@driveblocks.ai">driveblocks</maintainer>
@@ -15,7 +15,7 @@
   <depend>db_msgs</depend>
   <depend>geometry_msgs</depend>
   <depend>lanelet2_core</depend>
-  <depend>lib_mission_planner</depend>
+  <depend>autoware_local_mission_planner_common</depend>
   <depend>rclcpp</depend>
   <depend>tf2</depend>
   <depend>tf2_geometry_msgs</depend>
