+++ planning/mapless_architecture/autoware_local_mission_planner/src/mission_planner_node.cpp
@@ -1,11 +1,11 @@
 // Copyright 2024 driveblocks GmbH
 // driveblocks proprietary license
-#include "mission_planner_node.hpp"
+#include "autoware/local_mission_planner/mission_planner_node.hpp"
 
 #include "lanelet2_core/LaneletMap.h"
 #include "lanelet2_core/geometry/Lanelet.h"
 #include "lanelet2_core/geometry/LineString.h"
-#include "lib_mission_planner/helper_functions.hpp"
+#include "autoware/local_mission_planner_common/helper_functions.hpp"
 #include "autoware_planning_msgs/msg/driving_corridor.hpp"
 #include "autoware_planning_msgs/msg/mission.hpp"
 #include "autoware_planning_msgs/msg/mission_lanes_stamped.hpp"
