+++ planning/mapless_architecture/autoware_local_mission_planner/test/src/test_mission_planner_core.cpp
@@ -2,8 +2,8 @@
 // driveblocks proprietary license
 
 #include "gtest/gtest.h"
-#include "lib_mission_planner/helper_functions.hpp"
-#include "mission_planner_node.hpp"
+#include "autoware/local_mission_planner_common/helper_functions.hpp"
+#include "autoware/local_mission_planner/mission_planner_node.hpp"
 
 #include "db_msgs/msg/lanelets_stamped.hpp"
 #include "geometry_msgs/msg/pose.hpp"
