+++ planning/mapless_architecture/autoware_local_mission_planner_common/test/gtest_main.cpp
@@ -1,7 +1,6 @@
 // Copyright 2024 driveblocks GmbH
 // driveblocks proprietary license
 #include "gtest/gtest.h"
-#include "test_helper_functions.hpp"
 
 /**
  * Test helper functions
