+++ planning/mapless_architecture/autoware_mission_lane_converter/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mission_lane_converter</name>
+  <name>autoware_mission_lane_converter</name>
   <version>0.0.1</version>
   <description>Mission Lane Converter</description>
   <maintainer email="thomas.herrmann@driveblocks.ai">driveblocks</maintainer>
@@ -13,7 +13,7 @@
 
   <depend>autoware_auto_planning_msgs</depend>
   <depend>geometry_msgs</depend>
-  <depend>lib_mission_planner</depend>
+  <depend>autoware_local_mission_planner_common</depend>
   <depend>rclcpp</depend>
   <depend>tf2_geometry_msgs</depend>
   <depend>visualization_msgs</depend>
