+++ planning/mapless_architecture/autoware_mission_lane_converter/src/main_mission_lane_converter.cpp
@@ -1,6 +1,6 @@
 // Copyright 2024 driveblocks GmbH
 // driveblocks proprietary license
-#include "mission_lane_converter_node.hpp"
+#include "autoware/mission_lane_converter/mission_lane_converter_node.hpp"
 #include "rclcpp/rclcpp.hpp"
 
 #include <memory>
