--- planning/mapless_architecture/local_mission_planner_common/test/include/test_helper_functions.hpp
@@ -1,10 +1,0 @@
-// Copyright 2024 driveblocks GmbH
-// driveblocks proprietary license
-#ifndef TEST_HELPER_FUNCTIONS_HPP_
-#define TEST_HELPER_FUNCTIONS_HPP_
-
-#include "lib_mission_planner/helper_functions.hpp"
-
-// TODO(simon.eisenmann@driveblocks.ai): Add test functions
-
-#endif  // TEST_HELPER_FUNCTIONS_HPP_
