--- planning/mapless_architecture/local_mission_planner_common/test/src/test_helper_functions.cpp
@@ -1,8 +1,0 @@
-// Copyright 2024 driveblocks GmbH
-// driveblocks proprietary license
-#include "test_helper_functions.hpp"
-
-#include "gtest/gtest.h"
-#include "lib_mission_planner/helper_functions.hpp"
-
-// TODO(simon.eisenmann@driveblocks.ai): Add test functions
